use registerSharedProcessRemoteService to register
