+++ src/vs/workbench/services/userDataSync/electron-sandbox/userDataSyncStoreManagementService.ts
@@ -5,17 +5,16 @@
 
 import { IUserDataSyncStoreManagementService, UserDataSyncStoreType, IUserDataSyncStore } from 'vs/platform/userDataSync/common/userDataSync';
 import { ISharedProcessService } from 'vs/platform/ipc/electron-sandbox/services';
-import { IChannel } from 'vs/base/parts/ipc/common/ipc';
 import { IStorageService } from 'vs/platform/storage/common/storage';
 import { AbstractUserDataSyncStoreManagementService } from 'vs/platform/userDataSync/common/userDataSyncStoreService';
 import { IProductService } from 'vs/platform/product/common/productService';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
-import { URI } from 'vs/base/common/uri';
 import { registerSingleton } from 'vs/platform/instantiation/common/extensions';
+import { UserDataSyncStoreManagementServiceChannelClient } from 'vs/platform/userDataSync/common/userDataSyncIpc';
 
 class UserDataSyncStoreManagementService extends AbstractUserDataSyncStoreManagementService implements IUserDataSyncStoreManagementService {
 
-	private readonly channel: IChannel;
+	private readonly channelClient: UserDataSyncStoreManagementServiceChannelClient;
 
 	constructor(
 		@IProductService productService: IProductService,
@@ -24,28 +23,16 @@ class UserDataSyncStoreManagementService extends AbstractUserDataSyncStoreManage
 		@ISharedProcessService sharedProcessService: ISharedProcessService,
 	) {
 		super(productService, configurationService, storageService);
-		this.channel = sharedProcessService.getChannel('userDataSyncStoreManagement');
-		this._register(this.channel.listen('onDidChangeUserDataSyncStore')(() => this.updateUserDataSyncStore()));
+		this.channelClient = this._register(new UserDataSyncStoreManagementServiceChannelClient(sharedProcessService.getChannel('userDataSyncStoreManagement')));
+		this._register(this.channelClient.onDidChangeUserDataSyncStore(() => this.updateUserDataSyncStore()));
 	}
 
 	async switch(type: UserDataSyncStoreType): Promise<void> {
-		return this.channel.call('switch', [type]);
+		return this.channelClient.switch(type);
 	}
 
 	async getPreviousUserDataSyncStore(): Promise<IUserDataSyncStore> {
-		const userDataSyncStore = await this.channel.call<IUserDataSyncStore>('getPreviousUserDataSyncStore');
-		return this.revive(userDataSyncStore);
-	}
-
-	private revive(userDataSyncStore: IUserDataSyncStore): IUserDataSyncStore {
-		return {
-			url: URI.revive(userDataSyncStore.url),
-			defaultUrl: URI.revive(userDataSyncStore.defaultUrl),
-			insidersUrl: URI.revive(userDataSyncStore.insidersUrl),
-			stableUrl: URI.revive(userDataSyncStore.stableUrl),
-			canSwitch: userDataSyncStore.canSwitch,
-			authenticationProviders: userDataSyncStore.authenticationProviders,
-		};
+		return this.channelClient.getPreviousUserDataSyncStore();
 	}
 
 }
